refactor#97 mypage id type

+++ domain/mypage/src/main/java/com/teamsolply/solply/mypage/model/CourseInfoEntity.kt
@@ -3,7 +3,7 @@ package com.teamsolply.solply.mypage.model
 import com.teamsolply.solply.model.PlaceType
 
 data class CourseInfoEntity(
-    val courseId: Int,
+    val courseId: Long,
     val courseName: String,
     val placeTypeList: List<PlaceType>,
     val imageUrls: List<String>,
